* Divided current implementation into "trips.py" and "planned_trips.py". #22
